feat: Update 'service' and 'model' in 'src/direct_post' and 'src/creator'
This commit updates various codes in two directories.

In 'service.rs' from 'src/creator/' and 'src/direct_post/', methods for creating a new instance of the Service with token from environment variables are added. Also, methods for creating an instance with provided token and for setting custom base URL are updated.

In 'model.rs' from 'src/direct_post', new features such as enums for PostMode and MediaType are added as well as data structure for PhotoInitRequest.

In 'service.rs' from 'src/direct_post', new asynchronous function for initializing a photo post on TikTok and a simplified function to upload a photo from URLs that combines the steps of initializing a photo post and checking the post status are added.